Formulation/requirement.txt
asgiref==3.7.2
django-ckeditor==6.7.0
django-js-asset==2.1.0
sqlparse==0.4.4
typing-extensions==4.8.0
django-apscheduler==0.6.2
defusedxml==0.7.1
diff-match-patch==20230430
django-import-export==3.3.1
et-xmlfile==1.1.0
markuppy==1.14
odfpy==1.4.1
openpyxl==3.1.2
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
djangorestframework==3.14.0
defusedxml==0.7.1
diff-match-patch==20230430
django-import-export==3.3.1
et-xmlfile==1.1.0
markuppy==1.14
odfpy==1.4.1
openpyxl==3.1.2
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
django-simpleui==2023.8.28
Django==4.2.5
pulp==2.7.0
reportlab==4.0.5
pillow==10.0.1
django-cors-headers==4.2.0
numpy==1.26.0
pandas==2.1.1
python-dateutil==2.8.2
tzdata==2023.3
XlsxWriter==3.1.6
reportlab==4.0.5
requests==2.25.1
matplotlib==3.8.0
Pillow==10.0.1
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
kiwisolver==1.4.5
numpy==1.26.0
packaging==23.2
pyparsing==2.4.7
python-dateutil==2.8.2
